Fix: report 저장할때 존재하는 버그 수정

+++ src/main/java/GDSCKNU/CitySavior/repository/report/ReportRepository.java
@@ -17,7 +17,7 @@ public interface ReportRepository extends JpaRepository<Report, Long> {
     @Query(value = "SELECT * FROM Report r " +
             "WHERE ST_DWithin(r.location, :point, :radius, false) = true", nativeQuery = true)
     List<Report> findReportsWithinRadius(@Param("point") Point point,
-                                         @Param("radius") double radius);
+                                         @Param("radius") int radius);
 
     int countReportByMemberEmail(String email);
